+++ scm-core/src/main/java/sonia/scm/repository/spi/BranchStaleComputer.java
@@ -22,42 +22,17 @@
  * SOFTWARE.
  */
 
-package sonia.scm.plugin;
+package sonia.scm.repository.spi;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.Data;
+import sonia.scm.repository.Branch;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
+public interface BranchStaleComputer {
 
-class PendingPluginInstallation {
+  boolean computeStale(Branch branch, StaleContext context);
 
-  private static final Logger LOG = LoggerFactory.getLogger(PendingPluginInstallation.class);
-
-  private final AvailablePlugin plugin;
-  private final Path file;
-
-  PendingPluginInstallation(AvailablePlugin plugin, Path file) {
-    this.plugin = plugin;
-    this.file = file;
-  }
-
-  public AvailablePlugin getPlugin() {
-    return plugin;
-  }
-
-  void cancel() {
-    String name = plugin.getDescriptor().getInformation().getName();
-    LOG.info("cancel installation of plugin {}", name);
-    if (Files.exists(file)) {
-      try {
-        Files.delete(file);
-      } catch (IOException ex) {
-        throw new PluginFailedToCancelInstallationException("failed to cancel plugin installation ", name, ex);
-      }
-    } else {
-      LOG.info("plugin file {} did not exists for plugin {}; nothing deleted", file, name);
-    }
+  @Data
+  class StaleContext {
+    private Branch defaultBranch;
   }
 }