+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Add tooltips to short links on repository overview ([#1441](https://github.com/scm-manager/scm-manager/pull/1441))
+- Show the date of the last commit for branches in the frontend ([#1439](https://github.com/scm-manager/scm-manager/pull/1439))
+- Unify and add description to key view across user settings ([#1440](https://github.com/scm-manager/scm-manager/pull/1440))
+
+## [2.10.1] - 2020-11-24
+### Fixed
+- Improved logging of failures during plugin installation ([#1442](https://github.com/scm-manager/scm-manager/pull/1442))
+- Do not throw exception when plugin file does not exist on cancelled installation ([#1442](https://github.com/scm-manager/scm-manager/pull/1442))
+
 ## [2.10.0] - 2020-11-20
 ### Added
 - Delete branches directly in the UI ([#1422](https://github.com/scm-manager/scm-manager/pull/1422))
@@ -419,3 +430,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [2.9.0]: https://www.scm-manager.org/download/2.9.0
 [2.9.1]: https://www.scm-manager.org/download/2.9.1
 [2.10.0]: https://www.scm-manager.org/download/2.10.0
+[2.10.1]: https://www.scm-manager.org/download/2.10.1